+++ lib/light_paint.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 
 class LightPaint extends CustomPainter {
   final double progress;
@@ -10,9 +9,13 @@ class LightPaint extends CustomPainter {
 
   Paint _paintFocus;
 
-  LightPaint(this.progress, this.positioned, this.sizeCircle,
-      {this.colorShadow = Colors.black, this.opacityShadow = 0.8})
-      : assert(opacityShadow >= 0 && opacityShadow <= 1) {
+  LightPaint(
+    this.progress,
+    this.positioned,
+    this.sizeCircle, {
+    this.colorShadow = Colors.black,
+    this.opacityShadow = 0.8,
+  }) : assert(opacityShadow >= 0 && opacityShadow <= 1) {
     _paintFocus = Paint()
       ..color = Colors.transparent
       ..blendMode = BlendMode.clear;
@@ -23,9 +26,9 @@ class LightPaint extends CustomPainter {
     canvas.saveLayer(Offset.zero & size, Paint());
     canvas.drawColor(colorShadow.withOpacity(opacityShadow), BlendMode.dstATop);
 
-    var maxSise = size.width > size.height ? size.width : size.height;
+    var maxSize = size.width > size.height ? size.width : size.height;
 
-    double radius = maxSise * (1 - progress) + sizeCircle;
+    double radius = maxSize * (1 - progress) + sizeCircle;
 
     canvas.drawCircle(positioned, radius, _paintFocus);
     canvas.restore();